refactor: move themes to libs/facets
