+++ src/components/organisms/NoticeDetails/Feedbacks.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { Theme } from 'app/theme';
+import { Theme } from 'libs/facets/theme';
 import { Button } from 'components/atoms';
 
 interface FeedbacksProps {
